requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==1.0.15
click==8.2.1
coloredlogs==15.0.1
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
distro==1.9.0
durationpy==0.10
executing==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.3
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.1.4
langchain-community==0.0.38
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.8
lark==1.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.98.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.33
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==2.7.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.15.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.39.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
